requirements.txt
Pillow>=9.0.0
requests>=2.25.0
beautifulsoup4>=4.9.0
streamlit>=1.20.0